# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DeFarm Engines API
  version: 1.0.0
  description: |
    DeFarm Engines API provides a comprehensive agricultural supply chain tracking system
    with tokenization, circuit management, and secure item verification.

    ## Key Features
    - **JWT Authentication** - Secure token-based authentication
    - **Circuit Management** - Permission-controlled data sharing repositories
    - **Item Tokenization** - Local ID (LID) to DFID conversion via circuits
    - **Enhanced Identifiers** - Canonical and contextual identifier support
    - **Storage Adapters** - IPFS, Stellar blockchain, and custom storage backends
    - **Webhook System** - Post-action webhooks for circuit events
    - **Audit Trail** - Complete event history and activity tracking

    ## Authentication
    All authenticated endpoints require a JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    Or API key authentication:
    ```
    X-API-Key: dfm_<your-api-key>
    ```
  contact:
    name: DeFarm API Support
    email: api@defarm.net
  license:
    name: Proprietary
servers:
  - url: https://connect.defarm.net/api
    description: Production server
  - url: http://localhost:3000/api
    description: Local development server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key in format dfm_<32-character-string>
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        event_type:
          type: string
          enum:
            - ItemCreated
            - ItemEnriched
            - ItemMerged
            - ItemSplit
            - CircuitPush
            - CircuitPull
        dfid:
          type: string
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
        visibility:
          type: string
          enum: [Public, Private, CircuitOnly]
        is_encrypted:
          type: boolean
    LocalEventResponse:
      type: object
      properties:
        event_id:
          type: string
          format: uuid
        local_event_id:
          type: string
          format: uuid
          description: The local event ID for tracking before push
        event_type:
          type: string
          enum: [Created, Enriched, Transferred, Merged, Split, Circuit, Published, Verified, Custom]
        timestamp:
          type: integer
          format: int64
          description: Unix epoch timestamp
        source:
          type: string
          description: User ID who created the event
        metadata:
          type: object
          additionalProperties: true
        visibility:
          type: string
          enum: [Public, Private, CircuitOnly]
        is_local:
          type: boolean
          description: True if event has not been pushed to a circuit
...
    LogEventRequest:
      type: object
      required:
        - user_id
        - event_type
        - action
        - resource
        - outcome
        - severity
      properties:
        user_id:
          type: string
        event_type:
          type: string
          enum: [security, data, access, compliance, system, user]
        action:
          type: string
        resource:
          type: string
        resource_id:
          type: string
          nullable: true
        outcome:
          type: string
          enum: [success, failure, warning, blocked]
        severity:
          type: string
          enum: [low, medium, high, critical]
        details:
          type: object
          additionalProperties: true
        metadata:
          $ref: '#/components/schemas/AuditEventMetadataRequest'
        compliance:
          $ref: '#/components/schemas/ComplianceInfoRequest'
    AuditEventMetadataRequest:
      type: object
      properties:
        user_agent:
          type: string
        ip_address:
          type: string
        location:
          type: string
        device_id:
          type: string
        session_duration:
          type: integer
          format: int64
    ComplianceInfoRequest:
      type: object
      properties:
        gdpr:
          type: boolean
        ccpa:
          type: boolean
        hipaa:
          type: boolean
        sox:
          type: boolean
    AuditEventResponse:
      type: object
      properties:
        event_id:
          type: string
          format: uuid
        user_id:
          type: string
        event_type:
          type: string
        action:
          type: string
        resource:
          type: string
        resource_id:
          type: string
          nullable: true
        outcome:
          type: string
        severity:
          type: string
        timestamp:
          type: integer
          format: int64
        details:
          type: object
          additionalProperties: true
        metadata:
          $ref: '#/components/schemas/AuditEventMetadataResponse'
        signature:
          type: string
          nullable: true
        compliance:
          $ref: '#/components/schemas/ComplianceInfoResponse'
    AuditEventMetadataResponse:
      type: object
      properties:
        user_agent:
          type: string
        ip_address:
          type: string
        location:
          type: string
        device_id:
          type: string
        session_duration:
          type: integer
          format: int64
    ComplianceInfoResponse:
      type: object
      properties:
        gdpr:
          type: boolean
        ccpa:
          type: boolean
        hipaa:
          type: boolean
        sox:
          type: boolean
...
paths:
...
  /events:
    get:
      tags:
        - Events
      summary: List events
      description: Get event history with optional filters
      operationId: listEvents
      parameters:
        - name: dfid
          in: query
          description: Filter by DFID
          schema:
            type: string
        - name: event_type
          in: query
          description: Filter by event type
          schema:
            type: string
        - name: from_time
          in: query
          description: Start timestamp (Unix epoch)
          schema:
            type: integer
            format: int64
        - name: to_time
          in: query
          description: End timestamp (Unix epoch)
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
  /events/local:
    post:
      tags:
        - Events
      summary: Create local event
      description: |
        Create a local event that is not yet associated with a DFID.
        Local events can be pushed to a circuit later to associate them with items.
        Supports both JWT and API key authentication.
      operationId: createLocalEvent
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - event_type
                - visibility
              properties:
                event_type:
                  type: string
                  description: Type of event
                  enum: [Created, Enriched, Transferred, Merged, Split, Circuit, Published, Verified, Custom]
                  example: Created
                visibility:
                  type: string
                  description: Event visibility
                  enum: [Public, Private, CircuitOnly]
                  example: Public
                metadata:
                  type: object
                  additionalProperties: true
                  description: Additional metadata for the event
                  example:
                    action: initial_registration
                    location: Farm A
      responses:
        '200':
          description: Local event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocalEventResponse'
        '401':
          description: Authentication required
        '400':
          description: Invalid event type or visibility
  /events/local/{local_event_id}:
    get:
      tags:
        - Events
      summary: Get local event
      description: Retrieve a local event by its local_event_id
      operationId: getLocalEvent
      parameters:
        - name: local_event_id
          in: path
          required: true
          description: The local event UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Local event found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocalEventResponse'
        '404':
          description: Local event not found
        '400':
          description: Invalid local_event_id format
...
  /audit/events:
    post:
      tags:
        - Audit
      summary: Log audit event
      description: Create a new audit log event
      operationId: logAuditEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogEventRequest'
      responses:
        '200':
          description: Event logged successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      event_id:
                        type: string
                        format: uuid
...
  /audit/events/{event_id}:
    get:
      tags:
        - Audit
      summary: Get audit event by ID
      description: Retrieve a specific audit event
      operationId: getAuditEventById
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Audit event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventResponse'
...